refactor and document makeResponse
