+++ libs/state-machine/src/index.ts
@@ -2,7 +2,6 @@ export * from './lib/stateMachineButtonPressed';
 export * from './lib/handleWizardRequest';
 export * from './lib/setupInitialState';
 export * from './lib/makeResponse';
-export * from './lib/convertButtons';
 export * from './schemas';
 
 // exports types from libs/state-machine/src/lib/types.ts
