--- libs/state-machine/src/lib/convertButtons.ts
@@ -1,29 +1,0 @@
-import type {
-	CurrentStepRouteResponse,
-	WizardButton,
-	WizardStepLayout,
-	WizardStepLayoutButton,
-} from './types';
-
-export const convertWizardStepLayoutButtonsToWizardButtons = (
-	layoutButtons: WizardStepLayout['buttons'],
-): CurrentStepRouteResponse['buttons'] => {
-	const convertButton = (
-		index: string,
-		input: WizardStepLayoutButton,
-	): WizardButton => {
-		return {
-			id: index,
-			label: input.label,
-			buttonType: input.buttonType,
-		};
-	};
-
-	const outputRecord: CurrentStepRouteResponse['buttons'] = {};
-
-	Object.entries(layoutButtons).forEach(([index, layoutButton]) => {
-		outputRecord[index] = convertButton(index, layoutButton);
-	});
-
-	return outputRecord;
-};
