+++ libs/state-machine/src/lib/makeResponse.ts
@@ -1,12 +1,44 @@
-import { convertWizardStepLayoutButtonsToWizardButtons } from './convertButtons';
 import type {
 	CurrentStepRouteRequest,
 	CurrentStepRouteResponse,
+	WizardButton,
 	WizardStepData,
 	WizardStepLayout,
+	WizardStepLayoutButton,
 } from './types';
 
-export const getResponseFromBodyAndStateAndWizardStepLayout = (
+const convertWizardStepLayoutButtonsToWizardButtons = (
+	layoutButtons: WizardStepLayout['buttons'],
+): CurrentStepRouteResponse['buttons'] => {
+	const convertButton = (
+		index: string,
+		input: WizardStepLayoutButton,
+	): WizardButton => {
+		return {
+			id: index,
+			label: input.label,
+			buttonType: input.buttonType,
+		};
+	};
+
+	const outputRecord: CurrentStepRouteResponse['buttons'] = {};
+
+	Object.entries(layoutButtons).forEach(([index, layoutButton]) => {
+		outputRecord[index] = convertButton(index, layoutButton);
+	});
+
+	return outputRecord;
+};
+
+/**
+ * Constuct a CurrentStepRouteResponse to return
+ * to the client after the request has been handled.
+ *
+ * Involves constructing the markdown message and converting
+ * the data about the buttons on the next step to the format
+ * required by the frontend.
+ */
+export const makeResponse = (
 	requestBody: CurrentStepRouteRequest,
 	state: WizardStepData,
 	nextWizardStepLayout: WizardStepLayout,
